improve code?
